Добавлена Pydantic моделей заказов, вспомогательной функции для получения заказа

+++ app/routers/cart.py
@@ -92,9 +92,13 @@ async def remove_item_from_cart(product_id: int, current_user: UserModel = Depen
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail="Товар в корзине не найден")
 
-    await db.delete(cart_item)
+    if cart_item.quantity > 1:
+        cart_item.quantity -= 1
+        db.add(cart_item)
+    else:
+        await db.delete(cart_item)
+
     await db.commit()
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 @router.delete("/", status_code=status.HTTP_204_NO_CONTENT)
